Move stats view to correct package
